requirements.txt

aiogram==2.25.1
openai>=1.0.0
python-dotenv
